Few updates to BEDParser. Herald (Logger) class is added. A bug with coordinate deserializer needs to be resolved.

+++ Di3/BEDParser/ParsedBED.cs
@@ -14,7 +14,7 @@ namespace BEDParser
     {
         public string fileName { set; get; }
         public string filePath { set; get; }
-        public UInt64 fileHashKey { set; get; }
+        public UInt32 fileHashKey { set; get; }
         public byte chrCount { set; get; }
         public int peaksCount { set; get; }
         public Dictionary<string, ChrStatistics> chrStatistics { set; get; }
